# Conflicts:
#	.gitignore
#	backend/pom.xml
#	backend/src/main/java/com/example/SocialMedia/SocialMediaApplication.java
#	backend/src/main/java/com/example/SocialMedia/config/AppConfig.java
#	backend/src/main/java/com/example/SocialMedia/config/SecurityConfig.java
#	backend/src/main/java/com/example/SocialMedia/controller/AuthController.java
#	backend/src/main/java/com/example/SocialMedia/controller/PostController.java
#	backend/src/main/java/com/example/SocialMedia/keys/ConversationMembersID.java
#	backend/src/main/java/com/example/SocialMedia/keys/PinnedMessagesID.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Comment.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Follow.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/InteractableItems.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Post.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/PostMedia.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/PostTag.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Reaction.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Report.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Role.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Share.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/Story.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/User.java
#	backend/src/main/java/com/example/SocialMedia/model/coredata_model/UserRole.java
#	backend/src/main/java/com/example/SocialMedia/model/messaging_model/Conversation.java
#	backend/src/main/java/com/example/SocialMedia/model/messaging_model/ConversationMember.java
#	backend/src/main/java/com/example/SocialMedia/model/messaging_model/MessageMedia.java
#	backend/src/main/java/com/example/SocialMedia/model/messaging_model/PinnedMessages.java
#	backend/src/main/java/com/example/SocialMedia/repository/PostRepository.java
#	backend/src/main/java/com/example/SocialMedia/repository/UserRepository.java
#	backend/src/main/java/com/example/SocialMedia/service/PostService.java
#	backend/src/main/java/com/example/SocialMedia/service/UserService.java
#	backend/src/main/java/com/example/SocialMedia/serviceImpl/PostServiceImpl.java
#	backend/src/main/resources/application.yaml
#	database/Schema CoreData.sql
#	database/Schema Messaging.sql
#	database/function&trigger.sql
#	database/seedDataCore.sql
#	database/seedDataMessaging.sql
#	database/test.sql

+++ backend/src/main/java/com/example/SocialMedia/keys/PinnedMessagesID.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 @Embeddable
 public class PinnedMessagesID implements Serializable {
     private int conversationID;
-    private int messageID;
+    private long messageID;
 
     @Override
     public boolean equals(Object obj) {
